smartfruit/requirements.txt
Django==5.1.6
djangorestframework==3.16.0
scikit-learn==1.5.1
numpy==2.1.1
requests==2.32.3
gunicorn==23.0.0